refactor(ui): extract combobox filtering logic into custom hook
