+++ packages/ui-components/src/components/ComboBox/hooks/index.ts
@@ -1,7 +1,9 @@
 export {
-  default as useComboBoxMappedOptions,
+  default as useComboBoxOptionMapping,
   type OptionValuesAndLabelsKey,
   type OptionValuesAndLabelsValue,
-} from "./useComboBoxMappedOptions"
+} from "./useComboBoxOptionMapping"
 
 export { default as useComboBoxFloating, type ComboBoxFloatingReturn } from "./useComboBoxFloating"
+
+export { default as useComboBoxOptionFiltering } from "./useComboBoxOptionFiltering"
